requirements.txt
torch>=1.9.0
torchvision
tensorboard
tqdm
hydra-core
numpy
pandas
opencv-python
omegaconf
matplotlib
seaborn
scipy
bootstrapped
gymnasium
moviepy
